finished ejercicio 4

+++ 4. Composicion en c++/lib/Rectangulo/Rectangulo.h
@@ -13,7 +13,7 @@ class Rectangulo {
     void imprimeEsq();
     Coordenada obtieneSupIzq();
     Coordenada obtieneInfDer();
-    int obtieneArea();
+    double obtieneArea();
 };
 
 #endif
